Create a "notification-display-service" daemon
"notification-display-service" is run by a systemd unit
named "notifications-display.service" ; if we are running
under a Wayland or X11 profile, it will display new
notifications under the form of customized popup windows,
and delete them once they are validated with "Ok".
Otherwise, it will just print the messages to stderr
without deleting them.

Change-Id: Id33b01d67df4d731f666c2b145262642ca21cdbc
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>
Author: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>
Author: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>